requirements.txt
fastapi
uvicorn[standard]
python-srtm
numpy